ARM: shmobile: lager: Add da9063 PMIC device node for system restart
Add a device node for the da9063 PMIC, with subnodes for rtc and wdt.
Regulator support is not yet included.

This allows the system to be restarted when the watchdog timer times
out, or when a system restart is requested.

Signed-off-by: Geert Uytterhoeven <geert+renesas@glider.be>
Signed-off-by: Simon Horman <horms+renesas@verge.net.au>
(cherry picked from commit 46dd8a809effdc7ebe6ec760e3e421d5ac2a40f1)
Signed-off-by: Simon Horman <horms+renesas@verge.net.au>