ARM: imx25: iomux: disable SION bit on all GPIO pads
Pulling down the DATA4-7 while SION is enabled causes eSDHC2 card
interrupt not to occur.  Disabling the SION bits on those pads fixes
the problem but there is no indication of this behavior, not in the RM
nor the CE.

We apply the following workarounds

 - Disable the SION bit on all GPIO pads by fixing the macros in the
   iomux-mx25.h, and
 - Use DR register, instead of PSR, to read the GPIO pin value

I might be better to drop the SION bit for all pads, but we don't have
time to validate the all function at the moment.  There is some usages
of the SION bit to workaround the errata exist on the other chips in
the i.mx familiy.  We'd better be careful on this.

Reviewed-by: Yasushi SHOJI <yashi@atmark-techno.com>
Reviewed-by: Daisuke Mizobuchi <mizo@atmark-techno.com>
Reviewed-by: Makoto Harada <makoto.harada@atmark-techno.com>