ARM: imx: armadillo4x0: fix CMD5 timeout error problem at AWL13 initialization
Enabled the power switch before eSDHC2 is initialized and SD_CLK is running.

There is a case that AWL13 doesn't respond to CMD5(IO_SEND_OP_COND) if SD_CLK is
active before SD power. This causes a timeout error at SDIO interface
initialization with the following message.

	mmc1: error -110 whilst initialising SDIO card

We have two options to workaround this.
1. Enable power switch before eSDHC2 initialization
2. Use SD_CLK clock gating

Both options can workaround this problem but #2 will cause another problem
which needs to be investigated more. So we use #1.

It is guaranteed that regulator drivers are probed before its client drivers
by using higher priority initcall.

Reviewed-by: Yasushi SHOJI <yashi@atmark-techno.com>
Reviewed-by: Takahiro Nakamura <takahiro.nakamura@atmark-techno.com>